resolve conflicts, import issues, and cleanup from rebase
