+++ src/test/mystery-encounter/mystery-encounter.test.ts
@@ -1,10 +1,10 @@
-import {afterEach, beforeAll, beforeEach, expect, describe, it, vi} from "vitest";
+import { afterEach, beforeAll, beforeEach, expect, describe, it, vi } from "vitest";
 import * as overrides from "../../overrides";
 import GameManager from "#app/test/utils/gameManager";
 import Phaser from "phaser";
-import {Species} from "#enums/species";
-import {MysteryEncounterPhase} from "#app/phases/mystery-encounter-phase";
-import {MysteryEncounterType} from "#enums/mystery-encounter-type";
+import { Species } from "#enums/species";
+import { MysteryEncounterPhase } from "#app/phases/mystery-encounter-phase";
+import { MysteryEncounterType } from "#enums/mystery-encounter-type";
 
 describe("Mystery Encounters", () => {
   let phaserGame: Phaser.Game;
@@ -29,12 +29,12 @@ describe("Mystery Encounters", () => {
     // Seed guarantees wild encounter to be replaced by ME
     vi.spyOn(game.scene, "resetSeed").mockImplementation(() => {
       game.scene.waveSeed = "test";
-      Phaser.Math.RND.sow([ game.scene.waveSeed ]);
+      Phaser.Math.RND.sow([game.scene.waveSeed]);
       game.scene.rngCounter = 0;
     });
   });
 
-  it("Spawns a mystery encounter", async() => {
+  it("Spawns a mystery encounter", async () => {
     await game.runToMysteryEncounter([
       Species.CHARIZARD,
       Species.VOLCARONA
@@ -44,7 +44,7 @@ describe("Mystery Encounters", () => {
     expect(game.scene.getCurrentPhase().constructor.name).toBe(MysteryEncounterPhase.name);
   });
 
-  it("", async() => {
+  it("", async () => {
     await game.runToMysteryEncounter([
       Species.CHARIZARD,
       Species.VOLCARONA
@@ -54,16 +54,16 @@ describe("Mystery Encounters", () => {
     expect(game.scene.getCurrentPhase().constructor.name).toBe(MysteryEncounterPhase.name);
   });
 
-  it("spawns mysterious challengers encounter", async() => {
+  it("spawns mysterious challengers encounter", async () => {
   });
 
-  it("spawns mysterious chest encounter", async() => {
+  it("spawns mysterious chest encounter", async () => {
   });
 
-  it("spawns dark deal encounter", async() => {
+  it("spawns dark deal encounter", async () => {
   });
 
-  it("spawns fight or flight encounter", async() => {
+  it("spawns fight or flight encounter", async () => {
   });
 });
