+++ src/test/utils/phaseInterceptor.ts
@@ -16,7 +16,7 @@ import {
   NewBattlePhase,
   NextEncounterPhase,
   PostSummonPhase,
-  SelectGenderPhase,
+  SelectGenderPhase, SelectModifierPhase,
   SelectStarterPhase,
   SelectTargetPhase,
   ShinySparklePhase,
@@ -33,11 +33,10 @@ import {
   UnavailablePhase,
   VictoryPhase
 } from "#app/phases";
-import UI, {Mode} from "#app/ui/ui";
-import {Phase} from "#app/phase";
+import UI, { Mode } from "#app/ui/ui";
+import { Phase } from "#app/phase";
 import ErrorInterceptor from "#app/test/utils/errorInterceptor";
-import {QuietFormChangePhase} from "#app/form-change-phase";
-import {SelectModifierPhase} from "#app/phases/select-modifier-phase";
+import { QuietFormChangePhase } from "#app/form-change-phase";
 import {
   MysteryEncounterBattlePhase,
   MysteryEncounterOptionSelectedPhase,
