+++ src/test/vitest.setup.ts
@@ -9,13 +9,13 @@ import { initMoves } from "#app/data/move";
 import { initPokemonPrevolutions } from "#app/data/pokemon-evolutions";
 import { initPokemonForms } from "#app/data/pokemon-forms";
 import { initSpecies } from "#app/data/pokemon-species";
-import { initAchievements } from "#app/system/achv.js";
-import { initVouchers } from "#app/system/voucher.js";
 import { initStatsKeys } from "#app/ui/game-stats-ui-handler";
 import { beforeAll, beforeEach, vi } from "vitest";
 import * as overrides from "#app/overrides";
-import { initMysteryEncounterDialogue } from "#app/data/mystery-encounters/dialogue/mystery-encounter-dialogue";
 import { initMysteryEncounters } from "#app/data/mystery-encounters/mystery-encounters";
+import { initMysteryEncounterDialogue } from "#app/data/mystery-encounters/mystery-encounter-dialogue";
+import { initVouchers } from "#app/system/voucher";
+import { initAchievements } from "#app/system/achv";
 
 initVouchers();
 initAchievements();
